Add files via upload
"In this project, we employed a variety of techniques and models to tackle the problem of fake news classification. Our approach included:

Feature Representation:

We used **Bag of Words (BoW) **to represent the text data in a simple and efficient way.
We also applied ***Term Frequency-Inverse Document Frequency (TF-IDF):**to capture the importance of words in the context of the entire dataset.
Additionally, we harnessed the power of **pre-trained GloVe **word embeddings to create word vectors that carry semantic meaning.